# server/docs/swagger.yaml
basePath: /
definitions:
...
  apidocs.CreateProductRequest:
    properties:
      active:
        example: true
        type: boolean
      description:
        example: Rich cocoa layers
        type: string
      name:
        example: Chocolate cake
        type: string
      price_cents:
        example: 450000
        type: integer
      slug:
        example: chocolate-cake
        type: string
    type: object
...
  apidocs.ErrorBody:
    properties:
      code:
        example: validation_error
        type: string
      message:
        example: invalid request
        type: string
    type: object
  apidocs.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/apidocs.ErrorBody'
    type: object
...
  apidocs.MessageResponse:
    properties:
      message:
        example: ok
        type: string
    type: object
...
  apidocs.PatchProductRequest:
    properties:
      active:
        type: boolean
      description:
        type: string
      name:
        type: string
      price_cents:
        type: integer
      slug:
        type: string
    type: object
  apidocs.PatchRoleRequest:
    properties:
      permissions:
        example:
        - products.read
        - orders.read
        items:
          type: string
        type: array
    type: object
...
  apidocs.ProductsListResponse:
    properties:
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
    type: object
...
  apidocs.RolesListResponse:
    properties:
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
    type: object
...
  models.Product:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/models.ProductImage'
        type: array
      name:
        type: string
      price_cents:
        type: integer
      slug:
        type: string
      updated_at:
        type: string
    type: object
  models.ProductImage:
    properties:
      created_at:
        type: string
      external_url:
        type: string
      id:
        type: integer
      path:
        type: string
      product_id:
        type: integer
      sort_order:
        type: integer
    type: object
  models.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      label:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.RolePermission'
        type: array
      updated_at:
        type: string
    type: object
  models.RolePermission:
    properties:
      permission:
        type: string
      role_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API for Tofex restaurant/confectionery chain (staff admin + public
    guest ordering).
  title: Tofex API
  version: "1.0"
paths:
...
  /api/v1/admin/products:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.ProductsListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List products (admin)
      tags:
      - admin-products
    post:
      consumes:
      - application/json
      parameters:
      - description: Product
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apidocs.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create product
      tags:
      - admin-products
  /api/v1/admin/products/{id}:
    delete:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: no content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - admin-products
    patch:
      consumes:
      - application/json
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apidocs.PatchProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - admin-products
  /api/v1/admin/products/{id}/images:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Image file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ProductImage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload product image
      tags:
      - admin-products
  /api/v1/admin/roles:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.RolesListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List roles
      tags:
      - admin-roles
  /api/v1/admin/roles/{id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Permission keys
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apidocs.PatchRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update role permissions
      tags:
      - admin-roles
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"